:recycle: 클래스 공통화 및 Dto 필드 순서 수정
